Improve form validation UX with ActiveRecord::Errors
